+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## todo-tree-0.5.0
+
+### Added
+- Added `workflow init` command to scaffold `.github/workflows/todo-tree.yml`.
+- Generated workflow template now pins `alexandretrotel/todo-tree-action@v1.0.3`.
+
+### Documentation
+- Documented GitHub Actions setup with `tt workflow init`.
+
 ## todo-tree-0.4.0
 
 ### Breaking Changes